requirements.txt
pandas==2.2.3
numpy==2.1.3
smartapi-python==1.5.5
PyYAML==6.0.1
python-dotenv==1.0.0pyotp==2.9.0
logzero==1.7.0
websocket-client==1.9.0
streamlit==1.40.2
plotly==5.24.1
APScheduler==3.10.4pyotp==2.9.0
logzero==1.7.0
websocket-client==1.9.0
